+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 1.4.4
   - Support GHC 9.10
+  - Add encoders for packed repeated fields that reverse
+    the order of the elements and/or apply a function to each
+    element of a Foldable container to produce the desired value.
+  - Add bytesIfNonempty, unsafeFromByteString, unsafeFromShortByteString.
 
 1.4.3
   - Support GHC 9.8